+++ src/score/engine/obj-extension-line.ts
@@ -75,8 +75,12 @@ export class ObjExtensionLine extends MusicObject {
 
         if (obj instanceof ObjRhythmColumn) {
             const mcols = obj.measure.getColumns();
-            if (obj === mcols[0])
-                return obj.measure.getRect().left;
+            if (obj === mcols[0]) {
+                if (obj.measure.isFirstMeasureInRow())
+                    return obj.measure.getColumnsContentRect().left;
+                else
+                    return obj.measure.getRect().left;
+            }
         }
 
         return obj.getRect().right;